+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Fixed installation of Poetry git dependencies with a build system.
 - Fixed possible errors when resolving dependencies for specific packages.
+- Fixed handling of Python versions compatibility.
 
 
 ## [0.12.4] - 2018-10-21